+++ sasta/src/file_server/file_server.rs
@@ -1,16 +1,24 @@
-use std::{sync::{Mutex, Arc}, path::Path};
+use std::{
+    path::Path,
+    sync::{Arc, Mutex},
+};
 
-use axum::{extract::{State, Multipart}, Json, response::IntoResponse};
+use axum::{
+    body::Body,
+    extract::{Multipart, State},
+    response::IntoResponse,
+    Json,
+};
 use axum_macros::debug_handler;
-use hyper::{StatusCode, Request, Uri, Body};
+use hyper::{Request, StatusCode, Uri};
 use redis::{aio::Connection, Client, JsonAsyncCommands};
 use regex::Regex;
-use serde::{Serialize, Deserialize};
-use tokio::{sync::Mutex as AsyncMutex, io::AsyncWriteExt, fs::File as TokioFile};
+use serde::{Deserialize, Serialize};
+use tokio::{fs::File as TokioFile, io::AsyncWriteExt, sync::Mutex as AsyncMutex};
 use tokio_util::bytes::Bytes;
 use tower::ServiceExt;
 use tower_http::services::ServeFile;
-use tracing::{error_span, info_span, warn_span, warn};
+use tracing::{error_span, info_span, warn, warn_span};
 use uuid::Uuid;
 
 use crate::AppState;
@@ -21,10 +29,7 @@ pub type Response<T> = Result<Json<T>, (StatusCode, Json<T>)>;
 #[serde(tag = "type", content = "content")]
 pub enum Payload {
     FilePaths(Vec<TreeView>),
-    Error {
-        code: u8,
-        message: String
-    }
+    Error { code: u8, message: String },
 }
 
 #[derive(Serialize, Debug)]
@@ -32,23 +37,42 @@ pub struct TreeView {
     content: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     /// children is Some if content is a dir, None if content is a file
-    children: Option<Vec<TreeView>>
+    children: Option<Vec<TreeView>>,
 }
 
 impl From<&File> for TreeView {
     fn from(value: &File) -> Self {
-        TreeView { content: value.name.clone(), children: None }
+        TreeView {
+            content: value.name.clone(),
+            children: None,
+        }
     }
 }
 
 impl From<&Directory> for TreeView {
     fn from(value: &Directory) -> Self {
         let mut children = Vec::new();
-        children.append(&mut value.children.lock().unwrap().iter().map(|f| f.into()).collect::<Vec<TreeView>>());
-        children.append(&mut value.files.lock().unwrap().iter().map(|f| f.into()).collect::<Vec<TreeView>>());
+        children.append(
+            &mut value
+                .children
+                .lock()
+                .unwrap()
+                .iter()
+                .map(|f| f.into())
+                .collect::<Vec<TreeView>>(),
+        );
+        children.append(
+            &mut value
+                .files
+                .lock()
+                .unwrap()
+                .iter()
+                .map(|f| f.into())
+                .collect::<Vec<TreeView>>(),
+        );
         TreeView {
             content: value.name.clone(),
-            children: Some(children)
+            children: Some(children),
         }
     }
 }
@@ -60,22 +84,21 @@ pub async fn get_all_paths(State(state): State<AppState>) -> Response<Payload> {
 }
 
 #[debug_handler]
-pub async fn get_file(
-    State(state): State<AppState>,
-    uri: Uri
-) -> impl IntoResponse {
+pub async fn get_file(State(state): State<AppState>, uri: Uri) -> impl IntoResponse {
     let file_server = state.file_server.lock().await;
     let path = file_server.get_file(uri.to_string()).await;
 
     match path {
         Some(p) => {
-            let req = Request::builder().uri(uri.clone()).body(Body::empty()).unwrap();
+            let req = Request::builder()
+                .uri(uri.clone())
+                .body(Body::empty())
+                .unwrap();
             let f = ServeFile::new(format!("file_server/{p}"));
-            match f.oneshot(req).await {
-                Ok(res) => Ok(res.map(axum::body::boxed)),
-                Err(e) => Err((StatusCode::INTERNAL_SERVER_ERROR, format!("{e}"))),
-            }
-        },
+            f.oneshot(req)
+                .await
+                .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")))
+        }
         None => Err((StatusCode::NOT_FOUND, format!("{uri} not found"))),
     }
 }
@@ -88,16 +111,15 @@ struct AddFiles {
 #[debug_handler]
 pub async fn add_files(
     State(state): State<AppState>,
-    mut multipart: Multipart
+    mut multipart: Multipart,
 ) -> Response<Payload> {
     let mut file_server = state.file_server.lock().await;
     let mut add_files = AddFiles {
         directory: String::new(),
-        files: Vec::new()
+        files: Vec::new(),
     };
 
     while let Some(field) = multipart.next_field().await.unwrap() {
-
         if let Some(filename) = field.file_name() {
             let filename = filename.to_string();
             let bytes = field.bytes().await.unwrap();
@@ -106,7 +128,9 @@ pub async fn add_files(
         } else if let Some(name) = field.name() {
             // res.append(&mut vec!["WOHOO".to_string(), name.to_string(), field.text().await.unwrap()]);
             if name == "directory" {
-                let directory = field.text().await
+                let directory = field
+                    .text()
+                    .await
                     .unwrap_or(String::new())
                     .trim()
                     .trim_end_matches("/")
@@ -121,35 +145,47 @@ pub async fn add_files(
     }
 
     if add_files.directory.is_empty() {
-        return Err((StatusCode::BAD_REQUEST, Json(Payload::Error {
-            code: 2,
-            message: format!("Directory cannot be empty")
-        })))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(Payload::Error {
+                code: 2,
+                message: format!("Directory cannot be empty"),
+            }),
+        ));
     }
 
     let mut files = Vec::with_capacity(add_files.files.len());
     for (filename, bytes) in add_files.files {
-        match file_server.add_file(format!("{}/{filename}", add_files.directory)).await {
+        match file_server
+            .add_file(format!("{}/{filename}", add_files.directory))
+            .await
+        {
             Ok(f) => files.push((f, bytes)),
-            Err(message) => return Err((StatusCode::BAD_REQUEST, Json(Payload::Error { code: 3, message }))),
+            Err(message) => {
+                return Err((
+                    StatusCode::BAD_REQUEST,
+                    Json(Payload::Error { code: 3, message }),
+                ))
+            }
         }
     }
 
     file_server.write().await;
 
     for (f, bytes) in files {
-        let mut file = TokioFile::create(format!("file_server/{}", f.file_server)).await.unwrap();
+        let mut file = TokioFile::create(format!("file_server/{}", f.file_server))
+            .await
+            .unwrap();
         file.write_all(&bytes).await.unwrap();
     }
 
     Ok(Json(Payload::FilePaths(vec![])))
 }
 
-
 #[derive(Serialize, Deserialize, Clone)]
 pub struct File {
     name: String,
-    file_server: String
+    file_server: String,
 }
 
 #[derive(Clone)]
@@ -167,10 +203,10 @@ struct DesDir {
 }
 
 impl<'de> Deserialize<'de> for Directory {
-
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
-        D: serde::Deserializer<'de> {
+        D: serde::Deserializer<'de>,
+    {
         let input = DesDir::deserialize(deserializer)?;
         Ok(Self {
             name: input.name,
@@ -183,14 +219,15 @@ impl<'de> Deserialize<'de> for Directory {
 impl Serialize for Directory {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
-        S: serde::Serializer {
+        S: serde::Serializer,
+    {
         DesDir::serialize(
             &DesDir {
-                    name: self.name.clone(),
-                    files: self.files.lock().unwrap().to_vec(),
-                    children: self.children.lock().unwrap().to_vec()
-                },
-            serializer
+                name: self.name.clone(),
+                files: self.files.lock().unwrap().to_vec(),
+                children: self.children.lock().unwrap().to_vec(),
+            },
+            serializer,
         )
     }
 }
@@ -206,20 +243,21 @@ impl FileServer {
         let mut con = client.get_async_connection().await.unwrap();
 
         let root = match con.json_get::<_, _, String>("files", ".").await {
-            Ok(str) => {
-                serde_json::from_str(&str).unwrap()
-            },
+            Ok(str) => serde_json::from_str(&str).unwrap(),
             Err(e) => {
                 warn!("Could not parse files content, starting with a blank root directory (Error: {:?})", e);
                 Directory {
                     name: "".to_string(),
                     files: Arc::new(Mutex::new(vec![])),
-                    children: Arc::new(Mutex::new(vec![]))
+                    children: Arc::new(Mutex::new(vec![])),
                 }
-            },
+            }
         };
 
-        Self { con: AsyncMutex::new(con), root }
+        Self {
+            con: AsyncMutex::new(con),
+            root,
+        }
     }
 
     pub async fn get_paths(&self) -> TreeView {
@@ -234,7 +272,7 @@ impl FileServer {
         let file_path = file_path.trim();
         let re = Regex::new(r"^/[\w/_\-\.]+[\w]$").unwrap();
         if !re.is_match(file_path) {
-            return Err("Illegal file name. Must only contain '_-./' special characters, start with root ('/') and end with a letter.".to_string())
+            return Err("Illegal file name. Must only contain '_-./' special characters, start with root ('/') and end with a letter.".to_string());
         }
 
         let mut path = file_path
@@ -247,17 +285,23 @@ impl FileServer {
 
         let mut files = dir.files.lock().unwrap();
         match files.binary_search_by_key(&file_name, |f| &f.name) {
-            Ok(_) =>
-                Err(format!("File {file_path} already exists")),
+            Ok(_) => Err(format!("File {file_path} already exists")),
             Err(pos) => {
-                files.insert(pos, File {
-                    name: file_name.to_string(),
-                    file_server: format!("{}.{}", Uuid::new_v4(), Path::new(file_name).extension().unwrap().to_str().unwrap()),
-                    //TODO: Add things like path to file on disk (with uuid generated name)
-                });
+                files.insert(
+                    pos,
+                    File {
+                        name: file_name.to_string(),
+                        file_server: format!(
+                            "{}.{}",
+                            Uuid::new_v4(),
+                            Path::new(file_name).extension().unwrap().to_str().unwrap()
+                        ),
+                        //TODO: Add things like path to file on disk (with uuid generated name)
+                    },
+                );
 
                 Ok(files[pos].clone())
-            },
+            }
         }
     }
 
@@ -272,10 +316,8 @@ impl FileServer {
 
         let files = dir.files.lock().unwrap();
         match files.binary_search_by_key(&file_name, |f| &f.name) {
-            Ok(pos) =>
-                Some(files[pos].file_server.clone()),
-            Err(_) =>
-                None,
+            Ok(pos) => Some(files[pos].file_server.clone()),
+            Err(_) => None,
         }
     }
 
@@ -289,13 +331,16 @@ impl FileServer {
                 Ok(pos) => pos,
                 // Add Dir at sorted position in Vec if not present
                 Err(pos) => {
-                    d.insert(pos, Directory {
-                        name: p.to_string(),
-                        files: Arc::new(Mutex::new(vec![])),
-                        children: Arc::new(Mutex::new(vec![])),
-                    });
+                    d.insert(
+                        pos,
+                        Directory {
+                            name: p.to_string(),
+                            files: Arc::new(Mutex::new(vec![])),
+                            children: Arc::new(Mutex::new(vec![])),
+                        },
+                    );
                     pos
-                },
+                }
             };
             dir = d.get(pos).unwrap().clone();
         }
@@ -304,7 +349,13 @@ impl FileServer {
 
     async fn write(&mut self) {
         let root_dir = &self.root.clone();
-        if let Err(error) = self.con.lock().await.json_set::<_, _, _, String>("files", "$", &root_dir).await {
+        if let Err(error) = self
+            .con
+            .lock()
+            .await
+            .json_set::<_, _, _, String>("files", "$", &root_dir)
+            .await
+        {
             error_span!("Redis Error", ?error);
             // error_span!("Logging current state instead", ?self.content);
         }