+++ sasta/src/main.rs
@@ -1,29 +1,40 @@
-use std::{net::SocketAddr, sync::Arc, vec, collections::HashSet, env};
+use std::{collections::HashSet, env, net::SocketAddr, sync::Arc, vec};
 
-use axum::{Router, routing::{get, post, put, delete}, extract::{WebSocketUpgrade, ConnectInfo, State, Path}, response::IntoResponse, Json, ServiceExt};
+use axum::{
+    extract::{ConnectInfo, Path, Request, State, WebSocketUpgrade},
+    response::IntoResponse,
+    routing::{delete, get, post, put},
+    Json, Router, ServiceExt,
+};
 use axum_macros::debug_handler;
 use dotenv::dotenv;
 use hyper::StatusCode;
 use store::store::Store;
 use tokio::sync::{oneshot, Mutex};
-use tower_http::normalize_path::NormalizePath;
-use tracing::{Level, error, info, info_span};
-use tracing_subscriber::{FmtSubscriber, fmt::format::FmtSpan};
+use tracing::{error, info, info_span, Level};
+use tracing_subscriber::{fmt::format::FmtSpan, FmtSubscriber};
 use utoipa::OpenApi;
 use utoipa_rapidoc::RapiDoc;
 use utoipa_redoc::{Redoc, Servable};
 use utoipa_swagger_ui::SwaggerUi;
 use uuid::Uuid;
 
-use crate::{connection::connection::client_connection, file_server::file_server::{get_all_paths, FileServer, add_files, get_file}, store::schedule};
+use crate::{
+    connection::connection::client_connection,
+    file_server::file_server::{add_files, get_all_paths, get_file, FileServer},
+    store::schedule,
+};
 
-mod store;
 mod connection;
 mod file_server;
+mod store;
 
 impl From<(u8, String)> for read::Payload {
     fn from(value: (u8, String)) -> Self {
-        read::Payload::Error { code: value.0, message: value.1 }
+        read::Payload::Error {
+            code: value.0,
+            message: value.1,
+        }
     }
 }
 
@@ -88,51 +99,59 @@ async fn main() {
 
     let app_state = AppState { store, file_server };
 
-    let app = NormalizePath::trim_trailing_slash(
-        Router::new()
+    let app = Router::new()
         .merge(SwaggerUi::new("/swagger-ui").url("/api-docs/openapi.json", ApiDoc::openapi()))
         .merge(Redoc::with_url("/redoc", ApiDoc::openapi()))
-            // There is no need to create `RapiDoc::with_openapi` because the OpenApi is served
-            // via SwaggerUi instead we only make rapidoc to point to the existing doc.
-            // .merge(RapiDoc::new("/api-docs/openapi.json").path("/rapidoc"))
-            // Alternative to above
-            .merge(RapiDoc::with_openapi("/api-docs/openapi2.json", ApiDoc::openapi()).path("/rapidoc"))
-            .nest("/api", Router::new()
-                .nest("/display", Router::new()
-                    .route("/", get(read_display))
-                    .route("/", post(create_display))
-                    .route("/:uuid", put(update_display))
-                    .route("/:uuid", delete(delete_display))
+        // There is no need to create `RapiDoc::with_openapi` because the OpenApi is served
+        // via SwaggerUi instead we only make rapidoc to point to the existing doc.
+        // .merge(RapiDoc::new("/api-docs/openapi.json").path("/rapidoc"))
+        // Alternative to above
+        .merge(RapiDoc::with_openapi("/api-docs/openapi2.json", ApiDoc::openapi()).path("/rapidoc"))
+        .nest(
+            "/api",
+            Router::new()
+                .nest(
+                    "/display",
+                    Router::new()
+                        .route("/", get(read_display))
+                        .route("/", post(create_display))
+                        .route("/:uuid", put(update_display))
+                        .route("/:uuid", delete(delete_display)),
                 )
-                .nest("/playlist", Router::new()
-                    .route("/", post(create_playlist))
-                    .route("/", get(read_playlist))
-                    .route("/:uuid", put(update_playlist))
-                    .route("/:uuid", delete(delete_playlist))
+                .nest(
+                    "/playlist",
+                    Router::new()
+                        .route("/", post(create_playlist))
+                        .route("/", get(read_playlist))
+                        .route("/:uuid", put(update_playlist))
+                        .route("/:uuid", delete(delete_playlist)),
                 )
-                .nest("/schedule", Router::new()
-                    .route("/", post(create_schedule))
-                    .route("/", get(read_schedule))
-                    .route("/:uuid", put(update_schedule))
-                    .route("/:uuid", delete(delete_schedule))
+                .nest(
+                    "/schedule",
+                    Router::new()
+                        .route("/", post(create_schedule))
+                        .route("/", get(read_schedule))
+                        .route("/:uuid", put(update_schedule))
+                        .route("/:uuid", delete(delete_schedule)),
                 )
-                .nest("/files", Router::new()
-                    .route("/", get(get_all_paths))
-                    .route("/", post(add_files))
-                )
-            )
-            .nest("/files", Router::new()
-                .fallback(get_file)
-            )
-            .route("/", get(ws_handler))
-            .with_state(app_state)
-    );
+                .nest(
+                    "/files",
+                    Router::new()
+                        .route("/", get(get_all_paths))
+                        .route("/", post(add_files)),
+                ),
+        )
+        .nest("/files", Router::new().fallback(get_file))
+        .route("/", get(ws_handler))
+        .with_state(app_state);
 
     let addr = SocketAddr::from(([0, 0, 0, 0], 8040));
     info!("listening on {}", addr);
-    axum::Server::bind(&addr)
-        .serve(app.into_make_service_with_connect_info::<SocketAddr>())
-        .await.unwrap();
+    let listener = tokio::net::TcpListener::bind(&addr).await.unwrap();
+
+    axum::serve(listener, ServiceExt::<Request>::into_make_service(app))
+        .await
+        .unwrap();
 }
 
 mod read {
@@ -143,7 +162,7 @@ mod read {
     use utoipa::ToSchema;
     use uuid::Uuid;
 
-    use crate::store::{store, schedule};
+    use crate::store::{schedule, store};
 
     pub type Response = Result<Json<Payload>, (StatusCode, Json<Payload>)>;
 
@@ -155,10 +174,7 @@ mod read {
         Display(Vec<Display>),
         Playlist(Vec<Playlist>),
         Schedule(Vec<Schedule>),
-        Error {
-            code: u8,
-            message: String
-        }
+        Error { code: u8, message: String },
     }
 
     #[derive(Serialize, ToSchema, TS)]
@@ -168,12 +184,16 @@ mod read {
         pub uuid: Uuid,
         pub name: String,
         #[ts(type = "string")]
-        pub schedule: Uuid
+        pub schedule: Uuid,
     }
 
     impl From<(Uuid, store::Display)> for Display {
         fn from((uuid, d): (Uuid, store::Display)) -> Self {
-            Self { uuid, name: d.name, schedule: d.schedule }
+            Self {
+                uuid,
+                name: d.name,
+                schedule: d.schedule,
+            }
         }
     }
 
@@ -188,7 +208,11 @@ mod read {
 
     impl From<(Uuid, store::Playlist)> for Playlist {
         fn from((uuid, p): (Uuid, store::Playlist)) -> Self {
-            Self { uuid, name: p.name, items: p.items }
+            Self {
+                uuid,
+                name: p.name,
+                items: p.items,
+            }
         }
     }
 
@@ -200,13 +224,18 @@ mod read {
         pub name: String,
         #[ts(type = "string")]
         pub playlist: Uuid,
-        pub scheduled: Option<Vec<schedule::ScheduledPlaylistInput>>
+        pub scheduled: Option<Vec<schedule::ScheduledPlaylistInput>>,
     }
 
     impl From<(Uuid, schedule::Schedule)> for Schedule {
         fn from((uuid, s): (Uuid, schedule::Schedule)) -> Self {
             let s = schedule::ScheduleInput::from(s);
-            Self { uuid, name: s.name, playlist: s.playlist, scheduled: s.scheduled }
+            Self {
+                uuid,
+                name: s.name,
+                playlist: s.playlist,
+                scheduled: s.scheduled,
+            }
         }
     }
 }
@@ -225,13 +254,13 @@ mod create {
         pub uuid: Option<Uuid>,
         pub name: String,
         #[ts(type = "string")]
-        pub schedule: Uuid
+        pub schedule: Uuid,
     }
 
     #[derive(Deserialize, TS)]
     #[ts(export, export_to = "api_bindings/create/", rename = "CreatePlaylist")]
     pub struct Playlist {
-        pub name: String
+        pub name: String,
     }
 
     #[derive(Deserialize, TS)]
@@ -249,8 +278,8 @@ mod update {
     use utoipa::ToSchema;
     use uuid::Uuid;
 
-    pub use crate::read::{Response, Payload};
-    use crate::store::{store, schedule};
+    pub use crate::read::{Payload, Response};
+    use crate::store::{schedule, store};
 
     #[derive(Deserialize, ToSchema, TS)]
     #[ts(export, export_to = "api_bindings/update/", rename = "UpdateDisplay")]
@@ -258,7 +287,7 @@ mod update {
     pub struct Display {
         pub name: String,
         #[ts(type = "string")]
-        pub schedule: Uuid
+        pub schedule: Uuid,
     }
 
     #[derive(Deserialize, ToSchema, TS)]
@@ -277,35 +306,58 @@ mod update {
         #[ts(type = "string")]
         pub playlist: Uuid,
         #[ts(optional)]
-        pub scheduled: Option<Vec<schedule::ScheduledPlaylistInput>>
+        pub scheduled: Option<Vec<schedule::ScheduledPlaylistInput>>,
     }
 }
 
 #[debug_handler]
-async fn create_display(State(state): State<AppState>, Json(disp): Json<create::Display>) -> read::Response {
+async fn create_display(
+    State(state): State<AppState>,
+    Json(disp): Json<create::Display>,
+) -> read::Response {
     info_span!("[Api] Creating Display", display = ?disp);
     let store = state.store;
     let read = store.read().await;
     if let Some((uuid, _)) = read.displays.iter().find(|(_, d)| d.name == disp.name) {
         error!("[Api] Name is already used by Display {}", uuid);
-        return Err((StatusCode::BAD_REQUEST,
-            Json((1, format!("Avoid using the name {} as it is already used by another display", disp.name)).into())
-        ))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(
+                (
+                    1,
+                    format!(
+                        "Avoid using the name {} as it is already used by another display",
+                        disp.name
+                    ),
+                )
+                    .into(),
+            ),
+        ));
     }
 
     if let Some(uuid) = disp.uuid {
         if read.displays.contains_key(&uuid) {
             error!("[Api] Uuid is already used by another Display");
-            return Err((StatusCode::BAD_REQUEST,
-                Json((2, format!("Avoid using the Uuid {} as it is already used by another display", disp.name)).into())
-            ))
+            return Err((
+                StatusCode::BAD_REQUEST,
+                Json(
+                    (
+                        2,
+                        format!(
+                            "Avoid using the Uuid {} as it is already used by another display",
+                            disp.name
+                        ),
+                    )
+                        .into(),
+                ),
+            ));
         }
     }
     drop(read);
 
     let uuid = match disp.uuid {
         Some(u) => u,
-        None => Uuid::new_v4()
+        None => Uuid::new_v4(),
     };
     info!("[Api] Using Uuid {uuid} for new Display");
 
@@ -315,22 +367,39 @@ async fn create_display(State(state): State<AppState>, Json(disp): Json<create::
         info!("[Api] Created Display {uuid}");
         Ok(Json(read::Payload::Display(vec![(uuid, d.clone()).into()])))
     } else {
-        error!("[Api] No Display with {} could be found while reading after write", uuid);
-        Err((StatusCode::INTERNAL_SERVER_ERROR,
-            Json((3, format!("Something went wrong with the creation")).into())
+        error!(
+            "[Api] No Display with {} could be found while reading after write",
+            uuid
+        );
+        Err((
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json((3, format!("Something went wrong with the creation")).into()),
         ))
     };
 }
 
-async fn create_playlist(State(state): State<AppState>, Json(playlist): Json<create::Playlist>) -> create::Response {
+async fn create_playlist(
+    State(state): State<AppState>,
+    Json(playlist): Json<create::Playlist>,
+) -> create::Response {
     info!("[Api] Creating Playlist with name {}", playlist.name);
     let store = state.store;
     let read = store.read().await;
     if let Some((uuid, _)) = read.playlists.iter().find(|(_, p)| p.name == playlist.name) {
         error!("[Api] Name is already used by Playlist {}", uuid);
-        return Err((StatusCode::BAD_REQUEST,
-            Json((1, format!("Avoid using the name {} as it is already used by another Playlist", playlist.name)).into())
-        ))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(
+                (
+                    1,
+                    format!(
+                        "Avoid using the name {} as it is already used by another Playlist",
+                        playlist.name
+                    ),
+                )
+                    .into(),
+            ),
+        ));
     }
 
     drop(read);
@@ -340,38 +409,59 @@ async fn create_playlist(State(state): State<AppState>, Json(playlist): Json<cre
 
     return if let Some(p) = store.read().await.playlists.get(&uuid) {
         info!("[Api] Created Playlist {uuid}");
-        Ok(Json(read::Payload::Playlist(vec![(uuid, p.clone()).into()])))
+        Ok(Json(read::Payload::Playlist(
+            vec![(uuid, p.clone()).into()],
+        )))
     } else {
         error!("[Api] No Playlist with {uuid} could be found while reading after write");
-        Err((StatusCode::INTERNAL_SERVER_ERROR,
-            Json((2, format!("Something went wrong with the creation")).into())
+        Err((
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json((2, format!("Something went wrong with the creation")).into()),
         ))
     };
 }
 
-async fn create_schedule(State(state): State<AppState>, Json(schedule): Json<create::Schedule>) -> create::Response {
+async fn create_schedule(
+    State(state): State<AppState>,
+    Json(schedule): Json<create::Schedule>,
+) -> create::Response {
     info!("[Api] Creating Schedule with name {}", schedule.name);
     let store = state.store;
     let read = store.read().await;
     if let Some((uuid, _)) = read.schedules.iter().find(|(_, s)| s.name == schedule.name) {
         error!("[Api] Name is already used by Schedule {}", uuid);
-        return Err((StatusCode::BAD_REQUEST,
-            Json((1, format!("Avoid using the name {} as it is already used by another Schedule", schedule.name)).into())
-        ))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(
+                (
+                    1,
+                    format!(
+                        "Avoid using the name {} as it is already used by another Schedule",
+                        schedule.name
+                    ),
+                )
+                    .into(),
+            ),
+        ));
     }
 
     drop(read);
     let uuid = Uuid::new_v4();
     info!("[Api] Generated Uuid {uuid} for new Schedule");
-    store.create_schedule(uuid, schedule.name, schedule.playlist).await;
+    store
+        .create_schedule(uuid, schedule.name, schedule.playlist)
+        .await;
 
     return if let Some(s) = store.read().await.schedules.get(&uuid) {
         info!("[Api] Created Schedule {uuid}");
-        Ok(Json(read::Payload::Schedule(vec![(uuid, s.clone()).into()])))
+        Ok(Json(read::Payload::Schedule(
+            vec![(uuid, s.clone()).into()],
+        )))
     } else {
         error!("[Api] No Schedule with {uuid} could be found while reading after write");
-        Err((StatusCode::INTERNAL_SERVER_ERROR,
-            Json((2, format!("Something went wrong with the creation")).into())
+        Err((
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json((2, format!("Something went wrong with the creation")).into()),
         ))
     };
 }
@@ -394,13 +484,14 @@ async fn create_schedule(State(state): State<AppState>, Json(schedule): Json<cre
 )]
 async fn read_display(State(state): State<AppState>) -> read::Response {
     return Ok(Json(read::Payload::Display(
-            state.store
-                .read()
-                .await
-                .displays
-                .iter()
-                .map(|(u, d)| (*u, d.clone()).into())
-                .collect()
+        state
+            .store
+            .read()
+            .await
+            .displays
+            .iter()
+            .map(|(u, d)| (*u, d.clone()).into())
+            .collect(),
     )));
 }
 
@@ -429,13 +520,14 @@ async fn read_display(State(state): State<AppState>) -> read::Response {
 )]
 async fn read_playlist(State(state): State<AppState>) -> read::Response {
     return Ok(Json(read::Payload::Playlist(
-        state.store
+        state
+            .store
             .read()
             .await
             .playlists
             .iter()
             .map(|(u, p)| (*u, p.clone()).into())
-            .collect::<Vec<_>>()
+            .collect::<Vec<_>>(),
     )));
 }
 
@@ -462,13 +554,14 @@ async fn read_playlist(State(state): State<AppState>) -> read::Response {
 )]
 async fn read_schedule(State(state): State<AppState>) -> read::Response {
     return Ok(Json(read::Payload::Schedule(
-        state.store
+        state
+            .store
             .read()
             .await
             .schedules
             .iter()
             .map(|(u, s)| (*u, s.clone()).into())
-            .collect::<Vec<_>>()
+            .collect::<Vec<_>>(),
     )));
 }
 
@@ -504,68 +597,163 @@ async fn read_schedule(State(state): State<AppState>) -> read::Response {
         ("uuid" = Uuid, Path, description = "Uuid of Display to delete")
     )
 )]
-async fn update_display(State(state): State<AppState>, Path(uuid): Path<Uuid>, Json(display): Json<update::Display>) -> update::Response {
+async fn update_display(
+    State(state): State<AppState>,
+    Path(uuid): Path<Uuid>,
+    Json(display): Json<update::Display>,
+) -> update::Response {
     info!("[Api] Updating Display {uuid}");
     let store = state.store;
     let read = store.read().await;
     if !read.displays.contains_key(&uuid) {
         error!("[Api] No display with {uuid} was found");
-        return Err((StatusCode::BAD_REQUEST, Json((1, format!("No Display with the Uuid {uuid} was found")).into())))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json((1, format!("No Display with the Uuid {uuid} was found")).into()),
+        ));
     }
-    if let Some((uuid, _)) = read.displays.iter().find(|(&u, d)| d.name == display.name && u != uuid) {
+    if let Some((uuid, _)) = read
+        .displays
+        .iter()
+        .find(|(&u, d)| d.name == display.name && u != uuid)
+    {
         error!("[Api] Name is already used by Display {}", uuid);
-        return Err((StatusCode::BAD_REQUEST, Json((2, format!("Avoid using the name {} as it is already used by another display", display.name)).into())))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(
+                (
+                    2,
+                    format!(
+                        "Avoid using the name {} as it is already used by another display",
+                        display.name
+                    ),
+                )
+                    .into(),
+            ),
+        ));
     }
     drop(read);
 
-    store.update_display(uuid, display.name, display.schedule).await;
+    store
+        .update_display(uuid, display.name, display.schedule)
+        .await;
 
     return if let Some(d) = store.read().await.displays.get(&uuid) {
         info!("[Api] Updated and read Display {uuid}");
-        Ok(Json(update::Payload::Display(vec![(uuid, d.clone()).into()])))
+        Ok(Json(update::Payload::Display(vec![
+            (uuid, d.clone()).into()
+        ])))
     } else {
         error!("[Api] Could not find Display with {uuid} after update");
-        Err((StatusCode::INTERNAL_SERVER_ERROR, Json((3, format!("Could not find Display with {uuid} after update")).into())))
+        Err((
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(
+                (
+                    3,
+                    format!("Could not find Display with {uuid} after update"),
+                )
+                    .into(),
+            ),
+        ))
     };
 }
 
-async fn update_playlist(State(state): State<AppState>, Path(uuid): Path<Uuid>, Json(playlist): Json<update::Playlist>) -> update::Response {
+async fn update_playlist(
+    State(state): State<AppState>,
+    Path(uuid): Path<Uuid>,
+    Json(playlist): Json<update::Playlist>,
+) -> update::Response {
     info!("[Api] Updating Playlist {uuid}");
     let store = state.store;
     let read = store.read().await;
     if !read.playlists.contains_key(&uuid) {
         error!("[Api] No Playlist with {uuid} was found");
-        return Err((StatusCode::BAD_REQUEST, Json((1, format!("No Playlist with the Uuid {uuid} was found")).into())))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json((1, format!("No Playlist with the Uuid {uuid} was found")).into()),
+        ));
     }
-    if let Some((uuid, _)) = read.playlists.iter().find(|(&u, p)| p.name == playlist.name && u != uuid) {
+    if let Some((uuid, _)) = read
+        .playlists
+        .iter()
+        .find(|(&u, p)| p.name == playlist.name && u != uuid)
+    {
         error!("[Api] Name is already used by Playlist {}", uuid);
-        return Err((StatusCode::BAD_REQUEST, Json((2, format!("Avoid using the name {} as it is already used by another Playlist", playlist.name)).into())))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(
+                (
+                    2,
+                    format!(
+                        "Avoid using the name {} as it is already used by another Playlist",
+                        playlist.name
+                    ),
+                )
+                    .into(),
+            ),
+        ));
     }
 
     drop(read);
 
-    store.update_playlist(uuid, playlist.name, playlist.items).await;
+    store
+        .update_playlist(uuid, playlist.name, playlist.items)
+        .await;
 
     return if let Some(p) = store.read().await.playlists.get(&uuid) {
         info!("[Api] Updated and read Playlist {uuid}");
-        Ok(Json(update::Payload::Playlist(vec![(uuid, p.clone()).into()])))
+        Ok(Json(update::Payload::Playlist(vec![
+            (uuid, p.clone()).into()
+        ])))
     } else {
         error!("[Api] Could not find Playlist with {uuid} after update");
-        Err((StatusCode::INTERNAL_SERVER_ERROR, Json((4, format!("Could not find Playlist with {uuid} after update")).into())))
+        Err((
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(
+                (
+                    4,
+                    format!("Could not find Playlist with {uuid} after update"),
+                )
+                    .into(),
+            ),
+        ))
     };
 }
 
-async fn update_schedule(State(state): State<AppState>, Path(uuid): Path<Uuid>, Json(schedule): Json<update::Schedule>) -> update::Response {
+async fn update_schedule(
+    State(state): State<AppState>,
+    Path(uuid): Path<Uuid>,
+    Json(schedule): Json<update::Schedule>,
+) -> update::Response {
     info!("[Api] Updating Schedule {uuid}");
     let store = state.store;
     let read = store.read().await;
     if !read.schedules.contains_key(&uuid) {
         error!("[Api] No Schedule with {uuid} was found");
-        return Err((StatusCode::BAD_REQUEST, Json((1, format!("No Schedule with the Uuid {uuid} was found")).into())))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json((1, format!("No Schedule with the Uuid {uuid} was found")).into()),
+        ));
     }
-    if let Some((uuid, _)) = read.schedules.iter().find(|(&u, s)| s.name == schedule.name && u != uuid) {
+    if let Some((uuid, _)) = read
+        .schedules
+        .iter()
+        .find(|(&u, s)| s.name == schedule.name && u != uuid)
+    {
         error!("[Api] Name is already used by Schedule {}", uuid);
-        return Err((StatusCode::BAD_REQUEST, Json((2, format!("Avoid using the name {} as it is already used by another Schedule", schedule.name)).into())))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(
+                (
+                    2,
+                    format!(
+                        "Avoid using the name {} as it is already used by another Schedule",
+                        schedule.name
+                    ),
+                )
+                    .into(),
+            ),
+        ));
     }
 
     if let Some(scheduled) = &schedule.scheduled {
@@ -574,22 +762,46 @@ async fn update_schedule(State(state): State<AppState>, Path(uuid): Path<Uuid>, 
         // Checks if any playlist Uuid is a duplicate
         if !scheduled.iter().all(|s| uniq.insert(s.playlist)) {
             error!("[Api] Schedule contains duplicate Playlists");
-            return Err((StatusCode::BAD_REQUEST, Json((3, format!("Must not use the same Playlist more than once in a Schedule to avoid server meltdown")).into())))
+            return Err((StatusCode::BAD_REQUEST, Json((3, format!("Must not use the same Playlist more than once in a Schedule to avoid server meltdown")).into())));
         }
     }
     drop(read);
 
-    if let Err(e) = store.update_schedule(uuid, schedule.name, schedule.playlist, schedule.scheduled.unwrap_or(vec![])).await {
+    if let Err(e) = store
+        .update_schedule(
+            uuid,
+            schedule.name,
+            schedule.playlist,
+            schedule.scheduled.unwrap_or(vec![]),
+        )
+        .await
+    {
         error!("[Api] Schedule update failed with error: {e}");
-        return Err((StatusCode::INTERNAL_SERVER_ERROR, Json((4, format!("{e}")).into())))
+        return Err((
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json((4, format!("{e}")).into()),
+        ));
     }
 
     return if let Some(s) = store.read().await.schedules.get(&uuid) {
         info!("[Api] Updated and read Schedule {uuid}");
-        Ok(Json(update::Payload::Schedule(vec![(uuid, s.clone()).into()])))
+        Ok(Json(update::Payload::Schedule(vec![
+            (uuid, s.clone()).into()
+        ])))
     } else {
         error!("[Api] Could not find Schedule with {uuid} after update");
-        Err((StatusCode::INTERNAL_SERVER_ERROR, Json((5, format!("Could not find Schedule with {uuid} after update to avoid server meltdown")).into())))
+        Err((
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(
+                (
+                    5,
+                    format!(
+                        "Could not find Schedule with {uuid} after update to avoid server meltdown"
+                    ),
+                )
+                    .into(),
+            ),
+        ))
     };
 }
 
@@ -624,9 +836,10 @@ async fn delete_display(State(state): State<AppState>, Path(uuid): Path<Uuid>) -
         res = Ok(Json(read::Payload::Display(vec![(uuid, d.clone()).into()])));
     } else {
         error!("[Api] No display with {uuid} was found");
-        return Err((StatusCode::BAD_REQUEST,
-            Json((1, format!("No Display with the Uuid {uuid} was found")).into())
-        ))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json((1, format!("No Display with the Uuid {uuid} was found")).into()),
+        ));
     }
 
     store.delete_display(uuid).await;
@@ -665,35 +878,52 @@ async fn delete_display(State(state): State<AppState>, Path(uuid): Path<Uuid>) -
         ("uuid" = Uuid, Path, description = "Uuid of Playlist to delete")
     )
 )]
-pub(crate) async fn delete_playlist(State(state): State<AppState>, Path(uuid): Path<Uuid>) -> read::Response {
+pub(crate) async fn delete_playlist(
+    State(state): State<AppState>,
+    Path(uuid): Path<Uuid>,
+) -> read::Response {
     info!("[Api] Deleting Playlist {uuid}");
     let res;
     let store = state.store;
     let read = store.read().await;
 
-    let dependant_schedules = read.schedules.iter()
-        .filter_map(|(_, s)|
+    let dependant_schedules = read
+        .schedules
+        .iter()
+        .filter_map(|(_, s)| {
             if s.all_playlists().iter().any(|&p| p == &uuid) {
                 Some(s.name.clone())
             } else {
                 None
             }
-        ).collect::<Vec<_>>();
+        })
+        .collect::<Vec<_>>();
     if dependant_schedules.len() > 0 {
-        return Err((StatusCode::BAD_REQUEST,
-            Json((1, format!("Unable to delete playlist since the Schedules ({}) depend on it",
-                dependant_schedules.join(", ")
-            )).into())
-        ))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(
+                (
+                    1,
+                    format!(
+                        "Unable to delete playlist since the Schedules ({}) depend on it",
+                        dependant_schedules.join(", ")
+                    ),
+                )
+                    .into(),
+            ),
+        ));
     }
 
     if let Some(d) = read.playlists.get(&uuid) {
-        res = Ok(Json(read::Payload::Playlist(vec![(uuid, d.clone()).into()])));
+        res = Ok(Json(read::Payload::Playlist(
+            vec![(uuid, d.clone()).into()],
+        )));
     } else {
         error!("[Api] No Playlist with {uuid} was found");
-        return Err((StatusCode::BAD_REQUEST,
-            Json((2, format!("No Playlist with the Uuid {uuid} was found")).into())
-        ))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json((2, format!("No Playlist with the Uuid {uuid} was found")).into()),
+        ));
     }
 
     drop(read);
@@ -740,29 +970,43 @@ async fn delete_schedule(State(state): State<AppState>, Path(uuid): Path<Uuid>) 
     let store = state.store;
     let read = store.read().await;
 
-    let dependant_displays = read.displays.iter()
-        .filter_map(|(_, d)|
+    let dependant_displays = read
+        .displays
+        .iter()
+        .filter_map(|(_, d)| {
             if d.schedule == uuid {
                 Some(d.name.clone())
             } else {
                 None
             }
-        ).collect::<Vec<_>>();
+        })
+        .collect::<Vec<_>>();
     if dependant_displays.len() > 0 {
-        return Err((StatusCode::BAD_REQUEST,
-            Json((1, format!("Unable to delete playlist since the Displays ({}) depend on it",
-                dependant_displays.join(", ")
-            )).into())
-        ))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json(
+                (
+                    1,
+                    format!(
+                        "Unable to delete playlist since the Displays ({}) depend on it",
+                        dependant_displays.join(", ")
+                    ),
+                )
+                    .into(),
+            ),
+        ));
     }
 
     if let Some(s) = read.schedules.get(&uuid) {
-        res = Ok(Json(read::Payload::Schedule(vec![(uuid, s.clone()).into()])));
+        res = Ok(Json(read::Payload::Schedule(
+            vec![(uuid, s.clone()).into()],
+        )));
     } else {
         error!("[Api] No Schedule with {uuid} was found");
-        return Err((StatusCode::BAD_REQUEST,
-            Json((2, format!("No Schedule with the Uuid {uuid} was found")).into())
-        ))
+        return Err((
+            StatusCode::BAD_REQUEST,
+            Json((2, format!("No Schedule with the Uuid {uuid} was found")).into()),
+        ));
     }
 
     drop(read);
@@ -772,6 +1016,10 @@ async fn delete_schedule(State(state): State<AppState>, Path(uuid): Path<Uuid>) 
     res
 }
 
-async fn ws_handler(ws: WebSocketUpgrade, ConnectInfo(addr): ConnectInfo<SocketAddr>, State(state): State<AppState>) -> impl IntoResponse {
+async fn ws_handler(
+    ws: WebSocketUpgrade,
+    ConnectInfo(addr): ConnectInfo<SocketAddr>,
+    State(state): State<AppState>,
+) -> impl IntoResponse {
     ws.on_upgrade(move |socket| client_connection(socket, addr, state.store))
 }