+++ Changelog.md
@@ -1,4 +1,5 @@
  - A duration of zero now translates to an, effectively, infinite duration
  - Gasta container not displays Greet.md on login.
+ - Removed Sasta trim_trailing_slash() in path layer, meaning /api/test does not equal /api/test/ anymore due to causing swagger-ui infinite redirect error
 
 File mounted in Docker image as Greet.md will be displayed and rendered on login. Mounting a directory to /files/ allows static file hosting, which are inaccessible without login in.